LPS-55310 move tests to module
